tell the compiler I want these fast
git-svn-id: https://qthreads.googlecode.com/svn/trunk@3147 9dbd018f-d1da-d342-4c11-58c73dfe904b

+++ src/qt_mpool.c
@@ -77,9 +77,9 @@ typedef struct qt_mpool_cache_entry_s {
 
 struct threadlocal_cache_s {
     qt_mpool_cache_t             *cache;
-    size_t                        count;
+    uint_fast16_t                 count;
     uint8_t                      *block;
-    unsigned int                  i;
+    uint_fast32_t                 i;
     qt_mpool_threadlocal_cache_t *next;  // for cleanup
 };
